+++ LibTest/core/Set/remove_A01_t03.dart
@@ -23,7 +23,7 @@ class A implements Comparable {
   bool operator ==(dynamic other) => x == other.x;
   int get hashCode => -1;
   // to allow run this test for ordered sets we need A to be comparable
-  int compareTo(dynamic other) => this.x - other.x;
+  int compareTo(dynamic other) => this.x - other.x as int;
 }
 
 test(Set create([Set content])) {
